refactor(ci): split xtask ci task into srp modules (#0000)
